+++ R/p_value_kenward.R
@@ -75,7 +75,7 @@ p_value_kenward.lmerMod <- function(model, dof = NULL) {
   params <- insight::get_parameters(model, component = component)
 
   # check if all estimates are non-NA
-  params <- .check_rank_deficiency(params, verbose = FALSE)
+  params <- .check_rank_deficiency(model, params, verbose = FALSE)
 
   if (is.null(statistic)) {
     statistic <- insight::get_statistic(model, component = component)